requirements.txt
requests~=2.32.3
telebot~=0.0.5
peewee~=3.15.4
python-dotenv~=1.0.1